rename engine_store -> maintenance_engine_store, remove protocols in engine store creation
